+++ src/unix/notbsd/linux/other/b32/x86.rs
@@ -35,7 +35,7 @@ pub const SO_SNDTIMEO: ::c_int = 21;
 pub const FIOCLEX: ::c_ulong = 0x5451;
 pub const FIONBIO: ::c_ulong = 0x5421;
 
-pub const SYS_gettid: ::c_int = 224;
+pub const SYS_gettid: ::c_long = 224;
 
 extern {
     pub fn getcontext(ucp: *mut ucontext_t) -> ::c_int;
